+++ clients/Qt/oxrsys-home/src/MainWindow.h
@@ -110,6 +110,7 @@ private:
     QWidget* wiredHeadsetOptions_ = nullptr;
     QDoubleSpinBox* wiredEyeHeightSpin_ = nullptr;
     QCheckBox* wiredPositionTrackingCheckBox_ = nullptr;
+    QCheckBox* wiredControllerAdapterCheckBox_ = nullptr;
     QSpinBox* wiredDisplayIdSpin_ = nullptr;
     QComboBox* usbDeviceCombo_ = nullptr;
     QLabel* adbStatusLabel_ = nullptr;